backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
pandas==2.1.3
scikit-learn==1.3.2
numpy==1.26.2
mysqlclient==2.2.1
PyMySQL==1.1.0
